Cleanup printf format warnings in VGA S3

+++ src/hardware/vga_s3.cpp
@@ -350,7 +350,7 @@ void SVGA_S3_WriteCRTC(Bitu reg,Bitu val,Bitu /*iolen*/) {
 		vga.s3.reg_6b=(Bit8u)val;
 		break;
 	default:
-		LOG(LOG_VGAMISC,LOG_NORMAL)("VGA:S3:CRTC:Write to illegal index %2X", reg );
+		LOG(LOG_VGAMISC,LOG_NORMAL)("VGA:S3:CRTC:Write to illegal index %2X", static_cast<uint32_t>(reg));
 		break;
 	}
 }
@@ -469,7 +469,7 @@ void SVGA_S3_WriteSEQ(Bitu reg,Bitu val,Bitu /*iolen*/) {
 		VGA_StartResize();
 		break;
 	default:
-		LOG(LOG_VGAMISC,LOG_NORMAL)("VGA:S3:SEQ:Write to illegal index %2X", reg );
+		LOG(LOG_VGAMISC,LOG_NORMAL)("VGA:S3:SEQ:Write to illegal index %2X", static_cast<uint32_t>(reg));
 		break;
 	}
 }
@@ -494,7 +494,7 @@ Bitu SVGA_S3_ReadSEQ(Bitu reg,Bitu /*iolen*/) {
 	case 0x15:
 		return vga.s3.pll.cmd;
 	default:
-		LOG(LOG_VGAMISC,LOG_NORMAL)("VGA:S3:SEQ:Read from illegal index %2X", reg);
+		LOG(LOG_VGAMISC,LOG_NORMAL)("VGA:S3:SEQ:Read from illegal index %2X", static_cast<uint32_t>(reg));
 		return 0;
 	}
 }
